This commit introduces a faster _ParOr_. The algorithm is described informally in #140.

Benchmark data shows significant improvement compared to previous algorithm.

Benchmarks run on 15' 2017 MacBook Pro Intel(R) Core(TM) i7-7700HQ CPU @ 2.80GHz
```
benchmark                                           old ns/op     new ns/op     delta
BenchmarkRealDataParOr/census-income_srt-8          269673        194486        -27.88%
BenchmarkRealDataParOr/census-income-8              395797        310505        -21.55%
BenchmarkRealDataParOr/census1881_srt-8             723821        402359        -44.41%
BenchmarkRealDataParOr/census1881-8                 718247        573550        -20.15%
BenchmarkRealDataParOr/dimension_003-8              7466247       2197374       -70.57%
BenchmarkRealDataParOr/dimension_008-8              3062717       833368        -72.79%
BenchmarkRealDataParOr/dimension_033-8              474489        302080        -36.34%
BenchmarkRealDataParOr/uscensus2000-8               1912128       1092854       -42.85%
BenchmarkRealDataParOr/weather_sept_85_srt-8        370585        178577        -51.81%
BenchmarkRealDataParOr/weather_sept_85-8            1272372       1142709       -10.19%
BenchmarkRealDataParOr/wikileaks-noquotes_srt-8     290843        123167        -57.65%
BenchmarkRealDataParOr/wikileaks-noquotes-8         330745        152330        -53.94%
```

The previous algorithm is exposed in the public API as `ParHeapOr` somehow resembling the convention in `fastaggregation.go`.

Resolves #140

It's the second revision of the pull request. It fixed bugs found in #164

I've compared the cardinalities on /Real data/ dataset https://gist.github.com/maciej/add052d10c5edad2eb74eceda7484f07 semin-manually. They are the same, so I'm leaving #169 for later