@richardstartin 2018-04-10
Looks great, might be worth looking at uscensus2000 because it's quite an easy problem, the rest are comparable to the Java implementation (allowing for a different language and benchmarking machine).