@maciej 2018-04-10
@richardstartin could you link me to the Java ParOr benchmark results? I'd like to create a ticket regardin _uscensus2000_ dataset.

That said in the new `ParOr` there is still space for investigating performance improvements:
* the search for containers within a given key range is linear
* chunk size and count could be fine-tuned even more