@richardstartin 2018-04-10
There's a discussion (argument?) about it [here](https://github.com/RoaringBitmap/RoaringBitmap/issues/223), which only applies to the buffer implementation (typically the slower one). There's more discussion [https://github.com/RoaringBitmap/RoaringBitmap/pull/211](here) with various results (so look towards the end, not the beginning). Obviously, benchmarks depend on the machines they run on, but they are surprisingly similar to your results for several cases.

The obvious difference is that the Java implementation makes more effort to balance the tasks. This work takes anything between 10us and 400us. I think there is also a place in the Java project for an implementation assuming uniform key distribution as this one seems to, because it means you can sometimes get away with not doing the preparatory work.