@cakey 2018-04-10
Good and bad news. Your branch seems much much faster (100x?!), (awesome!) bad news is that our tests seem to get different results (off by 1 on some of the cardinalties)... Everything works if I switch back to ParHeapOr on your branch - Trying to work out if our code violates assumptions that don't cause problems in the old function, or if there might be a bug. Working on reproducing either way now.