@cakey 2018-04-10
Only seems to happen for parallelism > 1