@cakey 2018-04-11
@maciej 

```
	ba1 := []uint32{3585028, 65901253, 143441994, 211160474, 286511937, 356744840, 434332509, 502812785, 576097614, 646557334, 714794241, 775083485, 833704249, 889329147, 941367043}
	ba2 := []uint32{17883, 54494426, 113908938, 174519827, 235465665, 296685741, 357644666, 420192495, 476104304, 523046142, 577855081, 634889665, 692460635, 751350463, 809989192, 863494316, 919127240}

	r1 := roaring.BitmapOf(ba1...)
	r2 := roaring.BitmapOf(ba2...)

	result1 := roaring.ParOr(8, r1, r2)
	result2 := roaring.ParOr(1, r1, r2)
	result3 := roaring.ParHeapOr(1, r1, r2)
	result4 := roaring.ParHeapOr(8, r1, r2)

	fmt.Println(result1.GetCardinality())
	fmt.Println(result2.GetCardinality())
	fmt.Println(result3.GetCardinality())
	fmt.Println(result4.GetCardinality())
```
```
29
30
32
32
```

Possible