@lemire 2018-04-11
@maciej  I have not started on a code review, but given that we appear to have a reproducible failing test, maybe you want to have a look at that first?