@maciej 2018-04-12
@cakey the issue should now be fixed. Please rerun your tests.