@maciej 2018-04-13
@richardstartin if I interpret the [RoaringBitmap#211](https://github.com/RoaringBitmap/RoaringBitmap/pull/211) results correctly `parallelOr` on uscensus in Java takes somewhere around 200µs as compared to 1ms here?