@richardstartin 2018-04-13
@maciej Yes, but on a different machine, in a different language etc. I wrote an implementation that does uscens2000 in 90us but I dropped the PR. I think the critical difference is balanced task sizes.