@lemire 2018-04-13
@maciej Let me know when you recommend merging.