@maciej 2018-04-13
@lemire I've addressed your points from the review. As for `appendCopiesUntil` – you can only specify the stoppingKey without the start index. I decided not to introduce another method to the _roaringArray_ struct receiver. I think using the general `appendCopy` which abstract the CoW logic away is a better solution.

That said, once the CI checks are done I recommend merging.