@lemire 2018-04-13
The checks have failed, but it might be bogus. Restarting.