@lemire 2018-04-13
@cakey Care to comment on the speed-up of the new code on your internal tests?