@lemire 2018-04-14
This is now Version 0.4.5.

Send pizza to @maciej